fix avformat struct definitions
I removed all the fields that were not being used from the MpegTS* structs.  Then I looked at the differences between the struct definitions between what was committed and what was in the ffmpeg repo. For the difference that mattered I checked the git history and added a version check to include the field.

Futhermore, the MpegTSWriteStream struct supposedly contained a pointer to a MpegTSService struct, however, I could not find any version of ffmpeg where this field exists.  I removed it and removed the 2 references to it.

After these changes, I'm able to build picam and I no longer get a segfault in hls_write_packet when dereferencing ts_st->service (see a backtrace here https://github.com/iizukanao/picam/issues/144#issuecomment-690851912).