hls: fix EXT-X-TARGETDURATION
fix issue 152

According to the RFC (https://tools.ietf.org/html/rfc8216#section-4.3.3.1) EXT-X-TARGETDURATION specifies the "maximum Media Segment duration".  This changes the logic to find the maximum duration and use the nearest integer.  This also fixes a bug in the logic of calc_target_duration, the original author of this function had a copy/paste error where they forgot to include the logic to increment segment_durations_idx, so the code before was just calculating the average of the first value added to itself N times.